services/workflow-automation/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
PyYAML>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0